+++ mayan/apps/metadata/tests/test_workflow_actions.py
@@ -1,9 +1,15 @@
+from mayan.apps.document_states.events import event_workflow_edited
 from mayan.apps.document_states.permissions import permission_workflow_edit
 from mayan.apps.document_states.tests.mixins import (
     WorkflowTestMixin, WorkflowStateActionViewTestMixin
 )
 from mayan.apps.documents.tests.base import GenericDocumentViewTestCase
+from mayan.apps.events.tests.mixins import EventTestCaseMixin
 
+from ..events import (
+    event_document_metadata_added, event_document_metadata_edited,
+    event_document_metadata_removed
+)
 from ..models import MetadataType
 from ..workflow_actions import (
     DocumentMetadataAddAction, DocumentMetadataEditAction,
@@ -14,16 +20,18 @@ from ..permissions import (
     permission_document_metadata_remove
 )
 
-from .literals import TEST_METADATA_VALUE
+from .literals import (
+    DOCUMENT_METADATA_ADD_ACTION_CLASS_PATH,
+    DOCUMENT_METADATA_EDIT_ACTION_CLASS_PATH,
+    DOCUMENT_METADATA_REMOVE_ACTION_CLASS_PATH, TEST_METADATA_VALUE,
+    TEST_METADATA_VALUE_EDITED
+)
 from .mixins import DocumentMetadataMixin, MetadataTypeTestMixin
-
-DOCUMENT_METADATA_ADD_ACTION_CLASS_PATH = 'mayan.apps.metadata.workflow_actions.DocumentMetadataAddAction'
-DOCUMENT_METADATA_EDIT_ACTION_CLASS_PATH = 'mayan.apps.metadata.workflow_actions.DocumentMetadataEditAction'
-DOCUMENT_METADATA_REMOVE_ACTION_CLASS_PATH = 'mayan.apps.metadata.workflow_actions.DocumentMetadataRemoveAction'
 
 
 class DocumentMetadataActionTestCase(
-    DocumentMetadataMixin, MetadataTypeTestMixin
+    DocumentMetadataMixin, EventTestCaseMixin, MetadataTypeTestMixin,
+    GenericDocumentViewTestCase
 ):
     auto_upload_test_document = False
 
@@ -41,6 +49,9 @@ class DocumentMetadataActionTestCase(
         )
 
         metadata_count = self.test_document.metadata.count()
+
+        self._clear_events()
+
         action.execute(context={'document': self.test_document})
 
         self.assertEqual(
@@ -52,22 +63,41 @@ class DocumentMetadataActionTestCase(
             ).exists()
         )
 
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 1)
+
+        self.assertEqual(events[0].action_object, self.test_metadata_type)
+        self.assertEqual(events[0].actor, self.test_document)
+        self.assertEqual(events[0].target, self.test_document)
+        self.assertEqual(events[0].verb, event_document_metadata_added.id)
+
     def test_document_metadata_edit_action(self):
         self._create_test_document_metadata()
 
         action = DocumentMetadataEditAction(
             form_data={
                 'metadata_type': self.test_metadata_type.pk,
-                'value': TEST_METADATA_VALUE
+                'value': TEST_METADATA_VALUE_EDITED
             }
         )
 
         metadata_value = self.test_document.metadata.first().value
+
+        self._clear_events()
+
         action.execute(context={'document': self.test_document})
 
         self.assertNotEqual(
             metadata_value, self.test_document.metadata.first().value
         )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 1)
+
+        self.assertEqual(events[0].action_object, self.test_metadata_type)
+        self.assertEqual(events[0].actor, self.test_document)
+        self.assertEqual(events[0].target, self.test_document)
+        self.assertEqual(events[0].verb, event_document_metadata_edited.id)
 
     def test_document_metadata_remove_action(self):
         self._create_test_document_metadata()
@@ -77,6 +107,9 @@ class DocumentMetadataActionTestCase(
         )
 
         metadata_count = self.test_document.metadata.count()
+
+        self._clear_events()
+
         action.execute(context={'document': self.test_document})
 
         self.assertEqual(
@@ -88,9 +121,17 @@ class DocumentMetadataActionTestCase(
             ).exists()
         )
 
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 1)
+
+        self.assertEqual(events[0].action_object, self.test_metadata_type)
+        self.assertEqual(events[0].actor, self.test_document)
+        self.assertEqual(events[0].target, self.test_document)
+        self.assertEqual(events[0].verb, event_document_metadata_removed.id)
+
 
 class DocumentMetadataActionViewTestCase(
-    DocumentMetadataMixin, MetadataTypeTestMixin,
+    DocumentMetadataMixin, EventTestCaseMixin, MetadataTypeTestMixin,
     WorkflowStateActionViewTestMixin, WorkflowTestMixin,
     GenericDocumentViewTestCase
 ):
@@ -118,12 +159,25 @@ class DocumentMetadataActionViewTestCase(
             permission=permission_document_metadata_add
         )
 
+        self._clear_events()
+
         response = self._request_test_workflow_template_state_action_create_post_view(
             class_path=DOCUMENT_METADATA_ADD_ACTION_CLASS_PATH, extra_data={
                 'metadata_types': self.test_metadata_type.pk
             }
         )
         self.assertEqual(response.status_code, 302)
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 1)
+
+        self.assertEqual(
+            events[0].action_object,
+            self._test_workflow_template_state_action
+        )
+        self.assertEqual(events[0].actor, self._test_case_user)
+        self.assertEqual(events[0].target, self.test_workflow)
+        self.assertEqual(events[0].verb, event_workflow_edited.id)
 
     def test_document_metadata_edit_action_create_view(self):
         self._create_test_document_metadata()
@@ -136,6 +190,8 @@ class DocumentMetadataActionViewTestCase(
             permission=permission_document_metadata_edit
         )
 
+        self._clear_events()
+
         response = self._request_test_workflow_template_state_action_create_post_view(
             class_path=DOCUMENT_METADATA_EDIT_ACTION_CLASS_PATH, extra_data={
                 'metadata_type': self.test_metadata_type.pk,
@@ -143,6 +199,17 @@ class DocumentMetadataActionViewTestCase(
             }
         )
         self.assertEqual(response.status_code, 302)
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 1)
+
+        self.assertEqual(
+            events[0].action_object,
+            self._test_workflow_template_state_action
+        )
+        self.assertEqual(events[0].actor, self._test_case_user)
+        self.assertEqual(events[0].target, self.test_workflow)
+        self.assertEqual(events[0].verb, event_workflow_edited.id)
 
     def test_document_metadata_remove_action_create_view(self):
         self._create_test_document_metadata()
@@ -155,9 +222,23 @@ class DocumentMetadataActionViewTestCase(
             permission=permission_document_metadata_remove
         )
 
+        self._clear_events()
+
         response = self._request_test_workflow_template_state_action_create_post_view(
-            class_path=DOCUMENT_METADATA_REMOVE_ACTION_CLASS_PATH, extra_data={
+            class_path=DOCUMENT_METADATA_REMOVE_ACTION_CLASS_PATH,
+            extra_data={
                 'metadata_types': self.test_metadata_type.pk
             }
         )
         self.assertEqual(response.status_code, 302)
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 1)
+
+        self.assertEqual(
+            events[0].action_object,
+            self._test_workflow_template_state_action
+        )
+        self.assertEqual(events[0].actor, self._test_case_user)
+        self.assertEqual(events[0].target, self.test_workflow)
+        self.assertEqual(events[0].verb, event_workflow_edited.id)